Functionaliteit transaction
Exception moeten nog worden aangepast, en daylimit moet nog worden geimplementeerd

+++ src/main/java/io/swagger/model/NewTransaction.java
@@ -22,7 +22,7 @@ import javax.validation.constraints.*;
 @MappedSuperclass
 public class NewTransaction   {
   @JsonProperty("amount")
-  private BigDecimal amount = null;
+  private Long amount = null;
 
   @JsonProperty("timestamp")
   private LocalDateTime timestamp = null;
@@ -34,9 +34,9 @@ public class NewTransaction   {
   private String accountTo = null;
 
   @JsonProperty("performingUser")
-  private Integer performingUser = null;
+  private Long performingUser = null;
 
-  public NewTransaction amount(BigDecimal amount) {
+  public NewTransaction amount(Long amount) {
     this.amount = amount;
     return this;
   }
@@ -49,11 +49,11 @@ public class NewTransaction   {
       @NotNull
 
     @Valid
-    public BigDecimal getAmount() {
+    public Long getAmount() {
     return amount;
   }
 
-  public void setAmount(BigDecimal amount) {
+  public void setAmount(Long amount) {
     this.amount = amount;
   }
 
@@ -118,7 +118,7 @@ public class NewTransaction   {
     this.accountTo = accountTo;
   }
 
-  public NewTransaction performingUser(Integer performingUser) {
+  public NewTransaction performingUser(Long performingUser) {
     this.performingUser = performingUser;
     return this;
   }
@@ -130,11 +130,11 @@ public class NewTransaction   {
   @Schema(example = "1", required = true, description = "user id from the person who did the transaction (either customer or employee)")
       @NotNull
 
-    public Integer getPerformingUser() {
+    public Long getPerformingUser() {
     return performingUser;
   }
 
-  public void setPerformingUser(Integer performingUser) {
+  public void setPerformingUser(Long performingUser) {
     this.performingUser = performingUser;
   }
